requirements.txt
torch==1.4.0
torchvision==0.5.0
Pillow==7.2.0
tensorboardX==2.1
